build(deps): bump wasm-tools crates to 0.259 and wit-bindgen to 0.62
Move wasmparser, wasmprinter, wat, wasm-encoder, wit-component and
wasm-metadata to 0.259 and the wit-bindgen family to 0.62 so the toolchain
knows the component-model `external-id` name attribute. Enable the
`cm-implements` feature in the Wasm frontend validator, adapt the parser to
the `ComponentExternName` reader type and the `ComponentItem` validator
type, and follow the u64 offset and defined-type churn. The SDK macros
initialize the new `external_id` fields on wit-parser items.

+++ frontend/wasm/src/module/func_translator.rs
@@ -243,7 +243,7 @@ fn parse_function_body<B: ?Sized + Builder>(
         let (op, offset) = reader.read_with_offset().into_diagnostic()?;
         func_validator.op(pos, &op).into_diagnostic()?;
 
-        let dwarf_offset = module.wasm_file.dwarf_offset(offset as u64);
+        let dwarf_offset = module.wasm_file.dwarf_offset(offset);
         let span = resolve_instruction_span(addr2line, dwarf_offset, session, config)?;
         if !span.is_unknown() {
             last_valid_span = span;
